refactor: extract bundled extension manifest parser
(cherry picked from commit d307a7ca1a21b5650f74efa732368663c733a33d)
